+++ CHANGELOG.markdown
@@ -1,5 +1,6 @@
 # 0.1.2
 
+- [FIX] Don't split values containing `:` into broken pieces
 - Relax `rake` dependency
 - [DOC] Fill out README
 - Set up CI